ustabul-backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.8.1
django-filter==23.5
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
